src/frontend/src/app/services/fibonacci.service.ts
```typescript
import { Injectable } from '@angular/core';
import { Observable } from 'rxjs';

const TRUNCATE_AFTER_THIS_MANY = 10;

@Injectable({
  providedIn: 'root'
})
export class FibonacciService {

  validateInput(argument: string): boolean {
    // Validate user input as a valid integer
    return /^\d+$/.test(argument);
  }

  calculateFibonacci(argument: string): Observable<number[]> {
    // Generate Fibonacci numbers based on user input
    const num = parseInt(argument, 10);
    let fibList = [0, 1];
    for (let i = 2; i < num; i++) {
      fibList[i] = fibList[i - 1] + fibList[i - 2];
    }
    return new Observable<number[]>(observer => {
      observer.next(fibList);
      observer.complete();
    });
  }

  truncateFibonacciNumbers(numbers: number[]): number[] {
    // Truncate Fibonacci numbers if necessary
    return numbers.slice(0, TRUNCATE_AFTER_THIS_MANY);
  }

  callFibonacciAPI(argument: string): Observable<number[]> {
    // Communicate with backend API for Fibonacci calculation
    // Implement API call logic here
    return new Observable<number[]>(observer => {
      // Mock API response for testing
      observer.next([0, 1, 1, 2, 3, 5, 8, 13, 21, 34]);
      observer.complete();
    });
  }
}
```